requirements.txt
pytest~=8.4.1
selenium~=4.34.2
pytest-forked~=1.6.0
pytest-xdist~=3.8.0
pytest-html~=4.1.1
openpyxl~=3.1.5
allure-python-commons~=2.15.0